refactor(ports,adapters): define DiffLockfileReader port and extract parse_lockfile_content as shared pub fn
- Add `DiffSource` enum and `DiffLockfileReader` trait to new port file
  `src/ports/outbound/diff_lockfile_reader.rs`; register with
  `#[allow(unused_imports)]` re-export in `ports/outbound/mod.rs`
- Extract `parse_lockfile_content` and `parse_lockfile_content_for_member`
  from private `FileSystemReader` methods into public standalone functions in
  new `src/adapters/outbound/filesystem/lockfile_parser.rs`; module-scope
  shared structs (`UvDependency`, `DevDependencies`, `PackageSource`) and
  private helpers (`collect_all_deps`, `bfs_reachable`) reduce duplication
  and bring `parse_lockfile_content_for_member` well below the 30-line
  function threshold
- Update `FileSystemReader::read_and_parse_lockfile*` to call the free
  functions; remove now-redundant private methods and unused imports
- Update `LockfileReader` port doc comment to mention both `source.editable`
  and `source.virtual` (added in an earlier PR for uv >= 0.5 support)
- Add 9 unit tests to `lockfile_parser` (3 new + 6 migrated from
  `file_reader`); keep I/O integration test in `file_reader`

Closes #577
